# pkg/api/control-plane-openapi.yaml
...
openapi: 3.0.3
info:
  title: SRE Platform - Control Plane API
  description: |
    The Control Plane is the central management hub for the SRE Platform. It provides the Web UI and APIs for managing resources, users, incidents, and more. It orchestrates diagnostic tasks by calling the SRE Assistant service.
  version: 1.0.0
  contact:
    name: SRE Platform Team
    email: sre-platform@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8081
    description: Control Plane Development Server
  - url: https://api.sre-platform.example.com
    description: Production Server (Control Plane entry point)
...
paths:
...
  /api/v1/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: 儀表板摘要
      description: 獲取儀表板的摘要數據
      operationId: getDashboardSummary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 儀表板摘要
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /api/v1/dashboard/trends:
    get:
      tags: [Dashboard]
      summary: 趨勢數據
      description: 獲取各項指標的趨勢數據
      operationId: getDashboardTrends
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          description: 時間範圍
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: 24h
      responses:
        "200":
          description: 趨勢數據
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendData"
  /api/v1/dashboard/resource-distribution:
    get:
      tags: [Dashboard]
      summary: 資源分佈
      description: 獲取資源狀態分佈統計
      operationId: getResourceDistribution
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 資源分佈數據
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDistribution"

  # ============================================
  # Resources - 資源管理
  # ============================================
  /api/v1/resources:
    get:
      tags: [Resources]
      summary: 獲取資源列表
      description: 獲取所有資源的列表，支援分頁和過濾
      operationId: listResources
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          description: 資源狀態過濾
          schema:
            type: string
            enum: [healthy, warning, critical, unknown]
        - name: type
          in: query
          description: 資源類型過濾
          schema:
            type: string
            enum: [server, network, database, application, container]
        - name: group_id
          in: query
          description: 群組 ID 過濾
          schema:
            type: string
        - name: search
          in: query
          description: 搜尋關鍵字
          schema:
            type: string
      responses:
        "200":
          description: 資源列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceList"

    post:
      tags: [Resources]
      summary: 創建資源
      description: 創建新的資源
      operationId: createResource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceCreateRequest"
      responses:
        "201":
          description: 資源創建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/resources/{resourceId}:
    get:
      tags: [Resources]
      summary: 獲取資源詳情
      description: 獲取特定資源的詳細資訊
      operationId: getResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceIdParam"
      responses:
        "200":
          description: 資源詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Resources]
      summary: 更新資源
      description: 更新資源資訊
      operationId: updateResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdateRequest"
      responses:
        "200":
          description: 資源更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"

    delete:
      tags: [Resources]
      summary: 刪除資源
      description: 刪除資源
      operationId: deleteResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceIdParam"
      responses:
        "204":
          description: 資源刪除成功
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/resources/batch:
    post:
      tags: [Resources]
      summary: 批次操作資源
      description: 對多個資源執行批次操作
      operationId: batchOperateResources
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchResourceOperation"
      responses:
        "200":
          description: 批次操作結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchOperationResult"
  /api/v1/resources/scan:
    post:
      tags: [Resources]
      summary: 掃描網段
      description: 掃描指定網段發現新資源
      operationId: scanNetwork
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NetworkScanRequest"
      responses:
        "202":
          description: 掃描任務已接受
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanTaskResponse"
  /api/v1/resources/scan/{taskId}:
    get:
      tags: [Resources]
      summary: 獲取掃描結果
      description: 獲取網段掃描任務的結果
      operationId: getScanResult
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          description: 掃描任務 ID
          schema:
            type: string
      responses:
        "200":
          description: 掃描結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"

  # ============================================
  # Resource Groups - 資源群組管理
  # ============================================
...
  /api/v1/incidents:
    get:
      tags: [Incidents]
      summary: 獲取事件列表
      description: 獲取所有事件記錄
      operationId: listIncidents
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [new, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, error, warning, info]
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 事件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentList"

    post:
      tags: [Incidents]
      summary: 創建事件
      description: 創建新的事件記錄
      operationId: createIncident
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentCreateRequest"
      responses:
        "201":
          description: 事件創建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /api/v1/incidents/{incidentId}:
    get:
      tags: [Incidents]
      summary: 獲取事件詳情
      description: 獲取特定事件的詳細資訊
      operationId: getIncident
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 事件詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

    put:
      tags: [Incidents]
      summary: 更新事件
      description: 更新事件資訊
      operationId: updateIncident
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdateRequest"
      responses:
        "200":
          description: 事件更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /api/v1/incidents/{incidentId}/acknowledge:
    post:
      tags: [Incidents]
      summary: 確認事件
      description: 確認事件已被接收處理
      operationId: acknowledgeIncident
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                acknowledged_by:
                  type: string
                  description: 確認人
                comment:
                  type: string
                  description: 確認備註
      responses:
        "200":
          description: 事件確認成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /api/v1/incidents/{incidentId}/resolve:
    post:
      tags: [Incidents]
      summary: 解決事件
      description: 標記事件為已解決
      operationId: resolveIncident
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resolved_by:
                  type: string
                  description: 解決人
                resolution:
                  type: string
                  description: 解決方案描述
                root_cause:
                  type: string
                  description: 根本原因
      responses:
        "200":
          description: 事件解決成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /api/v1/incidents/{incidentId}/assign:
    post:
      tags: [Incidents]
      summary: 指派事件
      description: 指派事件給特定人員處理
      operationId: assignIncident
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assignee_id
              properties:
                assignee_id:
                  type: string
                  description: 被指派人 ID
      responses:
        "200":
          description: 事件指派成功
  /api/v1/incidents/{incidentId}/comments:
    post:
      tags: [Incidents]
      summary: 新增事件註記
      description: 為事件新增註記
      operationId: addIncidentComment
      security:
        - bearerAuth: []
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
                  description: 註記內容
      responses:
        "201":
          description: 註記新增成功
  /api/v1/incidents/generate-report:
    post:
      tags: [Incidents]
      summary: AI 生成事件報告
      description: 使用 AI 生成事件分析報告 (由 SRE Assistant 處理)
      operationId: generateIncidentReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - incident_ids
              properties:
                incident_ids:
                  type: array
                  items:
                    type: string
                  description: 要分析的事件 ID 列表
                report_type:
                  type: string
                  enum: [summary, root_cause, impact_analysis]
                  default: summary
      responses:
        "200":
          description: AI 報告生成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGeneratedReport"
  /api/v1/alerts:
    get:
      tags: [Incidents]
      summary: 獲取當前告警列表
      description: 獲取所有活躍的告警
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [P0, P1, P2, P3]
          description: 嚴重程度
        - name: status
          in: query
          schema:
            type: string
            enum: [firing, resolved, pending]
          description: 告警狀態
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 告警列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        severity:
                          type: string
                          enum: [P0, P1, P2, P3]
                        status:
                          type: string
                          enum: [firing, resolved, pending]
                        resource_name:
                          type: string
                        message:
                          type: string
                        fired_at:
                          type: string
                          format: date-time
                        resolved_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  firing_count:
                    type: integer
                  resolved_count:
                    type: integer

  # ============================================
  # Grafana Provisioning - Grafana 資源管理
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT Token
    grafanaApiToken:
      type: apiKey
      in: header
      name: X-Grafana-Authorization
      description: 用於授權存取 Grafana Provisioning API 的金鑰
  parameters:
    PageParam:
      name: page
      in: query
      description: 頁碼（從 1 開始）
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每頁筆數
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ResourceIdParam:
      name: resourceId
      in: path
      required: true
      description: 資源 ID
      schema:
        type: string
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: 當前頁碼
        page_size:
          type: integer
          description: 每頁筆數
        total:
          type: integer
          description: 總筆數
        total_pages:
          type: integer
          description: 總頁數

    # ============================================
    # Core Schemas
    # ============================================
...
    DashboardSummary:
      type: object
      properties:
        alerts:
          type: object
          properties:
            new:
              type: integer
            processing:
              type: integer
            resolved_today:
              type: integer
        resources:
          type: object
          properties:
            total:
              type: integer
            healthy:
              type: integer
            warning:
              type: integer
            critical:
              type: integer
        kpis:
          type: object
          properties:
            availability:
              type: number
              format: float
              description: 系統可用性百分比
            mttr:
              type: number
              description: 平均修復時間（分鐘）
            incident_rate:
              type: number
              description: 事件發生率
    TrendData:
      type: object
      properties:
        period:
          type: string
        data_points:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              alerts:
                type: integer
              incidents:
                type: integer
              resources_healthy:
                type: integer
              resources_unhealthy:
                type: integer
    ResourceDistribution:
      type: object
      properties:
        by_status:
          type: object
          properties:
            healthy:
              type: integer
            warning:
              type: integer
            critical:
              type: integer
            unknown:
              type: integer
        by_type:
          type: object
          additionalProperties:
            type: integer
        by_group:
          type: array
          items:
            type: object
            properties:
              group_id:
                type: string
              group_name:
                type: string
              total:
                type: integer
              healthy:
                type: integer
              warning:
                type: integer
              critical:
                type: integer

    # ============================================
    # Resource Schemas
    # ============================================
    Resource:
      type: object
      required:
        - id
        - name
        - type
        - status
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [server, network, database, application, container]
        status:
          type: string
          enum: [healthy, warning, critical, unknown]
        ip_address:
          type: string
        location:
          type: string
        group_id:
          type: string
        group_name:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
          description: 資源的鍵值對標籤
        owner:
          type: string
          description: 負責人
        description:
          type: string
        maintenance_window:
          $ref: "#/components/schemas/MaintenanceWindow"
        metrics:
          type: object
          properties:
            cpu_usage:
              type: number
              format: float
            memory_usage:
              type: number
              format: float
            disk_usage:
              type: number
              format: float
            network_io:
              type: number
              format: float
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ResourceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ResourceCreateRequest:
      type: object
      required:
        - name
        - type
        - ip_address
      properties:
        name:
          type: string
        type:
          type: string
          enum: [server, network, database, application, container]
        ip_address:
          type: string
        location:
          type: string
        group_id:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
          description: 資源的鍵值對標籤
        owner:
          type: string
        description:
          type: string
        metadata:
          type: object
    ResourceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        group_id:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
          description: 資源的鍵值對標籤
        owner:
          type: string
        description:
          type: string
        metadata:
          type: object
    BatchResourceOperation:
      type: object
      required:
        - operation
        - resource_ids
      properties:
        operation:
          type: string
          enum: [delete, add_to_group, remove_from_group, update_tags]
        resource_ids:
          type: array
          items:
            type: string
        parameters:
          type: object
          properties:
            group_id:
              type: string
            tags:
              type: object
              additionalProperties:
                type: string
              description: 要更新的鍵值對標籤
    BatchOperationResult:
      type: object
      properties:
        success_count:
          type: integer
        failure_count:
          type: integer
        failures:
          type: array
          items:
            type: object
            properties:
              resource_id:
                type: string
              error:
                type: string
    NetworkScanRequest:
      type: object
      required:
        - cidr
      properties:
        cidr:
          type: string
          description: CIDR 格式的網段（如 192.168.1.0/24）
        port_ranges:
          type: array
          items:
            type: string
          default: ["22", "80", "443", "3389"]
        timeout:
          type: integer
          description: 掃描超時時間（秒）
          default: 30
    ScanTaskResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, scanning, completed, failed]
        message:
          type: string
    ScanResult:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [completed, failed]
        discovered_resources:
          type: array
          items:
            type: object
            properties:
              ip_address:
                type: string
              hostname:
                type: string
              open_ports:
                type: array
                items:
                  type: integer
              os_detection:
                type: string
              services:
                type: array
                items:
                  type: string
        scan_time:
          type: string
          format: date-time
        error:
          type: string

    # ============================================
    # Resource Group Schemas
    # ============================================
...
    IncidentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        pagination:
          $ref: "#/components/schemas/Pagination"
    IncidentCreateRequest:
      type: object
      required:
        - title
        - severity
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, error, warning, info]
        affected_resources:
          type: array
          items:
            type: string
        tags:
          type: object
          additionalProperties:
            type: string
          description: 事件的鍵值對標籤
    IncidentUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, error, warning, info]
        affected_resources:
          type: array
          items:
            type: string
        tags:
          type: object
          additionalProperties:
            type: string
          description: 事件的鍵值對標籤
...